# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " Mojo Api",
        "description": "MOJO API Descrição",
        "contact": {
            "email": "wesleymotabr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://back.mojosolutions.co/api/",
            "description": "Server Mojo"
        }
    ],
    "paths": {
...
        "/content": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "Get Content",
                "description": "Returns list of contents",
                "operationId": "get_content",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "content"
                ],
                "summary": "Post de conteúdo",
                "description": "Cadastra um conteúdo",
                "operationId": "post_content",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "text"
                                    },
                                    "body": {
                                        "type": "text"
                                    },
                                    "link": {
                                        "type": "text"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "mamamia",
                                    "body": "blabal",
                                    "link": "",
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/content/{id}": {
            "put": {
                "tags": [
                    "content"
                ],
                "summary": "Put de conteúdo",
                "description": "Edita um conteúdo",
                "operationId": "put_content",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "text"
                                    },
                                    "body": {
                                        "type": "text"
                                    },
                                    "link": {
                                        "type": "text"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "mamamia",
                                    "body": "blabal",
                                    "link": "",
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "content"
                ],
                "summary": "Delete Content",
                "description": "Deleta um conteúdo",
                "operationId": "delete_content",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/password/find/{token}": {
            "get": {
                "tags": [
                    "reset_password"
                ],
                "summary": "Check if exist token",
                "description": "Returns list of Communities",
                "operationId": "check_token_password",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/password/create": {
            "post": {
                "tags": [
                    "reset_password"
                ],
                "summary": "Envia um email para resetar a senha",
                "description": "Envia um email para resetar a senha",
                "operationId": "reset_password_request",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "teste@gmail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "reset_password"
                ],
                "summary": "Form reset password",
                "description": "Formulário de resetar a senha",
                "operationId": "reset_password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    },
                                    "token": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "text"
                                    },
                                    "password_confirmation": {
                                        "type": "text"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "wesleymotabr@gmail.com",
                                    "token": "2UOdHfRrCp2qzHIy8eI15C6T2IR4Q9VgXkgOBqyARmnhzfKxD1T9rNaJXKLq",
                                    "password": "secret",
                                    "password_confirmation": "secret"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
}